backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.1.0
celery==5.4.0
httpx==0.27.0
beautifulsoup4==4.12.0
scrapy==2.11.0
playwright==1.47.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.5.0
xgboost==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pytest==8.3.0
pytest-asyncio==0.24.0
aiosqlite==0.20.0
python-socketio==5.11.0
pytz==2024.1
pytest-httpx==0.30.0
prometheus-fastapi-instrumentator==7.0.0
tqdm==4.66.0
